+++ includes/class-api.php
@@ -8,6 +8,7 @@
 
 namespace HRSWP\SQLSRV\API;
 use HRSWP\SQLSRV\Setup;
+use HRSWP\SQLSRV\Sqlsrv_Query;
 use HRSWP\SQLSRV\Sqlsrv_DB;
 
 if ( ! defined( 'ABSPATH' ) ) {
@@ -69,7 +70,47 @@ class API {
 				'methods'             => 'GET',
 				'callback'            => array( $this, 'get_tables_list' ),
 				'permission_callback' => function () {
-					return current_user_can( 'edit_others_posts' );
+					return current_user_can( 'edit_posts' );
+				},
+			)
+		);
+
+		// Register a public route to get job classification data.
+		register_rest_route(
+			$this->namespace,
+			'/jobclassification/table/(?P<table>[a-z0-9_\-]+)',
+			array(
+				'methods'             => 'GET',
+				'callback'            => array( $this, 'get_job_classification_data' ),
+				'args'                => array(
+					'table' => array(
+						'sanitize_callback' => function( $param ) {
+							return sanitize_key( $param );
+						},
+					),
+				),
+				'permission_callback' => function () {
+					return current_user_can( 'edit_posts' );
+				},
+			)
+		);
+
+		// Register a public route to get job classification data.
+		register_rest_route(
+			$this->namespace,
+			'/salarydata/table/(?P<table>[a-z0-9_\-]+)',
+			array(
+				'methods'             => 'GET',
+				'callback'            => array( $this, 'get_salary_data' ),
+				'args'                => array(
+					'table' => array(
+						'sanitize_callback' => function( $param ) {
+							return sanitize_key( $param );
+						},
+					),
+				),
+				'permission_callback' => function () {
+					return current_user_can( 'edit_posts' );
 				},
 			)
 		);
@@ -111,5 +152,80 @@ class API {
 		}
 
 		return new \WP_REST_Response( $tables, 200 );
+	}
+
+	/**
+	 * Returns data from a given Job Classification data table.
+	 *
+	 * @since 0.10.0
+	 *
+	 * @param \WP_REST_Request $request Data from the request passed to the callback.
+	 * @return array|null JSON feed of returned data, null if no data is found.
+	 */
+	public function get_job_classification_data( \WP_REST_Request $request ) {
+		if ( ! $request || '' === $request['table'] ) {
+			return new \WP_Error( 'missing-table', __( 'No table specified for query.', 'hrswp-sqlsrv-db' ) );
+		}
+
+		if ( 'undefined' !== $request['table'] ) {
+			$args = array(
+				'dataset' => array(
+					array(
+						'table'  => isset( $request['table'] ) ? sanitize_key( $request['table'] ) : '',
+						'fields' => array(
+							'ClassCode',
+							'JobTitle',
+							'SalRangeNum',
+							'SalrangeWExceptions',
+							'Salary_Min',
+							'Salary_Max',
+						),
+					),
+				),
+				'orderby' => 'JobTitle',
+			);
+
+			$result = new Sqlsrv_Query\Sqlsrv_Query( $args );
+			$result = ( ! empty( $result->records ) )
+				? $result->records
+				: array( 'request' => $request['table'] );
+
+			return new \WP_REST_Response( $result, 200 );
+		}
+
+		return new \WP_REST_Response( array( 'request' => $request['table'] ), 304 );
+	}
+
+	/**
+	 * Returns data from a given Salary data table.
+	 *
+	 * @since 0.10.0
+	 *
+	 * @param \WP_REST_Request $request Data from the request passed to the callback.
+	 * @return array|null JSON feed of returned data, null if no data is found.
+	 */
+	public function get_salary_data( \WP_REST_Request $request ) {
+		if ( ! $request || '' === $request['table'] ) {
+			return new \WP_Error( 'missing-table', __( 'No table specified for query.', 'hrswp-sqlsrv-db' ) );
+		}
+
+		if ( 'undefined' !== $request['table'] ) {
+			$args = array(
+				'dataset' => array(
+					array(
+						'table' => isset( $request['table'] ) ? sanitize_key( $request['table'] ) : '',
+					),
+				),
+			);
+
+			$result = new Sqlsrv_Query\Sqlsrv_Query( $args );
+			$result = ( ! empty( $result->records ) )
+				? $result->records
+				: array( 'request' => $request['table'] );
+
+			return new \WP_REST_Response( $result, 200 );
+		}
+
+		return new \WP_REST_Response( array( 'request' => $request['table'] ), 304 );
 	}
 }