+++ src/blocks/job-classifications/edit.js
@@ -1,10 +1,11 @@
 /**
+ * External dependencies
+ */
+import { escape, unescape } from 'lodash';
+
+/**
  * WordPress dependencies
  */
-const { __ } = wp.i18n;
-const { withSelect } = wp.data;
-const { InspectorControls } = wp.blockEditor;
-const ServerSideRender = wp.serverSideRender;
 const {
 	PanelBody,
 	Placeholder,
@@ -14,72 +15,211 @@ const {
 	TextControl,
 	ToggleControl,
 } = wp.components;
+const { dispatch, useSelect } = wp.data;
+const { InspectorControls, useBlockProps } = wp.blockEditor;
+const { __ } = wp.i18n;
 
-function JobClassifications( { attributes, setAttributes, tables } ) {
-	const {
-		className,
-		columns,
-		isStriped,
-		salaryDataUrl,
-		queryTable,
-	} = attributes;
+export default function JobClassificationsEdit( {
+	attributes: { displayAsList, columns, salaryDataUrl, queryTable },
+	setAttributes,
+} ) {
+	const { jobClassificationData, isRequesting, tables } = useSelect(
+		( select ) => {
+			const {
+				getJobClassificationData,
+				getTableNames,
+				isResolving,
+			} = select( 'hrswpsqlsrv/salary-data' );
 
-	const inspectorControls = (
-		<InspectorControls>
-			{ 'is-style-list' !== className && (
-				<PanelBody title={ __( 'Table Settings' ) }>
-					<ToggleControl
-						label={ __( 'Striped table rows' ) }
-						checked={ isStriped }
-						onChange={ ( value ) =>
-							setAttributes( { isStriped: value } )
-						}
-					/>
-				</PanelBody>
-			) }
-			{ 'is-style-list' === className && (
-				<PanelBody title={ __( 'List Settings' ) }>
-					<RangeControl
-						label={ __( 'List Columns' ) }
-						value={ columns || 3 }
-						onChange={ ( value ) =>
-							setAttributes( { columns: value } )
-						}
-						min={ 1 }
-						max={ 6 }
-					/>
-				</PanelBody>
-			) }
-			<PanelBody title={ __( 'Select Job Data Source' ) }>
-				<SelectControl
-					className={ 'salary-data-table-picker__select' }
-					label={ __( 'Select desired group:' ) }
-					value={ queryTable }
-					options={ tables }
-					onChange={ ( value ) =>
-						setAttributes( { queryTable: value } )
-					}
-				/>
-			</PanelBody>
-			<PanelBody title={ __( 'Salary Data Settings' ) }>
-				<TextControl
-					label={ __( 'Linked Salary Data URL' ) }
-					help={ __(
-						'The full URL to a page with a corresponding Salary Data block to link to. Leave blank to link to the current page.'
-					) }
-					value={ salaryDataUrl }
-					onChange={ ( value ) =>
-						setAttributes( { salaryDataUrl: value } )
-					}
-				/>
-			</PanelBody>
-		</InspectorControls>
+			return {
+				jobClassificationData: queryTable?.length
+					? getJobClassificationData( queryTable )
+					: {},
+				isRequesting: isResolving( 'getJobClassificationData', [
+					queryTable,
+				] ),
+				tables: getTableNames(),
+			};
+		},
+		[ queryTable ]
 	);
 
-	if ( ! queryTable ) {
+	const toggleAttribute = ( attributeName ) => ( newValue ) => {
+		if ( 'queryTable' === attributeName ) {
+			dispatch(
+				'hrswpsqlsrv/salary-data'
+			).invalidateResolutionForStoreSelector(
+				'getJobClassificationData'
+			);
+		}
+		setAttributes( { [ attributeName ]: newValue } );
+	};
+	const formatNumber = new Intl.NumberFormat( 'en-US', {
+		style: 'currency',
+		currency: 'USD',
+	} );
+
+	const getQueryTables = () => {
+		if ( ! tables?.length ) {
+			return [];
+		}
+		return tables?.reduce( ( accumulator, currentValue ) => {
+			if (
+				currentValue.value.includes( 'job-class' ) ||
+				'' === currentValue.value
+			) {
+				accumulator.push( currentValue );
+			}
+			return accumulator;
+		}, [] );
+	};
+
+	const renderJobClassificationName = ( name ) =>
+		! name ? __( '(Untitled)' ) : unescape( name ).trim();
+	const renderJobClassificationCurrency = ( number ) =>
+		! Number.isNaN( Number( number ) )
+			? formatNumber.format( number )
+			: renderJobClassificationName( number );
+	const renderJobClassificationRangeURL = ( rangeURLParam, range ) => {
+		const url = escape( salaryDataUrl + '?filter=' + rangeURLParam );
 		return (
-			<>
-				{ inspectorControls }
+			<a href={ url } target="_blank" rel="noreferrer noopener">
+				{ renderJobClassificationName( range ) }
+			</a>
+		);
+	};
+
+	const renderJobClassificationTable = () => {
+		return (
+			<figure className="wp-block-table">
+				<table>
+					<thead>
+						<tr>
+							<th>{ __( 'Job Class' ) }</th>
+							<th>{ __( 'Job Title' ) }</th>
+							<th>{ __( 'Range' ) }</th>
+							<th>{ __( 'Salary Min' ) }</th>
+							<th>{ __( 'Salary Max' ) }</th>
+						</tr>
+					</thead>
+					<tbody>
+						{ jobClassificationData.map(
+							( jobClassification, key ) =>
+								renderJobClassificationTableRow(
+									jobClassification,
+									key
+								)
+						) }
+					</tbody>
+				</table>
+			</figure>
+		);
+	};
+
+	const renderJobClassificationTableRow = ( jobClassification, key ) => {
+		const {
+			ClassCode: code,
+			JobTitle: title,
+			SalRangeNum: rangeURLParam,
+			Salary_Max: max,
+			Salary_Min: min,
+			SalrangeWExceptions: range,
+		} = jobClassification;
+
+		return (
+			<tr key={ key }>
+				<td>{ renderJobClassificationName( code ) }</td>
+				<td>{ renderJobClassificationName( title ) }</td>
+				<td>
+					{ renderJobClassificationRangeURL( rangeURLParam, range ) }
+				</td>
+				<td>{ renderJobClassificationCurrency( min ) }</td>
+				<td>{ renderJobClassificationCurrency( max ) }</td>
+			</tr>
+		);
+	};
+
+	const renderJobClassificationList = () => {
+		return (
+			<ul className={ `has-columns has-columns-${ columns }` }>
+				{ jobClassificationData.map( ( jobClassification, key ) =>
+					renderJobClassificationListItem( jobClassification, key )
+				) }
+			</ul>
+		);
+	};
+
+	const renderJobClassificationListItem = ( jobClassification, key ) => {
+		const {
+			ClassCode: code,
+			JobTitle: title,
+			SalRangeNum: rangeURLParam,
+			Salary_Max: max,
+			Salary_Min: min,
+			SalrangeWExceptions: range,
+		} = jobClassification;
+
+		return (
+			<li key={ key }>
+				<strong>{ renderJobClassificationName( title ) }</strong>
+				<span>{ ` (${ renderJobClassificationName( code ) })` }</span>
+				<ul>
+					<li>
+						{ ` ${ __( 'Range' ) }: ` }
+						{ renderJobClassificationRangeURL(
+							rangeURLParam,
+							range
+						) }
+					</li>
+					<li>
+						{ ` ${ __( 'Salary Min' ) }: ` }
+						{ renderJobClassificationCurrency( min ) }
+					</li>
+					<li>
+						{ ` ${ __( 'Salary Max' ) }: ` }
+						{ renderJobClassificationCurrency( max ) }
+					</li>
+				</ul>
+			</li>
+		);
+	};
+
+	return (
+		<div { ...useBlockProps() }>
+			<InspectorControls>
+				<PanelBody title={ __( 'Job Classifications settings' ) }>
+					<ToggleControl
+						label={ __( 'Display as list' ) }
+						checked={ displayAsList }
+						onChange={ toggleAttribute( 'displayAsList' ) }
+					/>
+					{ displayAsList && (
+						<RangeControl
+							label={ __( 'List Columns' ) }
+							value={ columns || 3 }
+							onChange={ toggleAttribute( 'columns' ) }
+							min={ 1 }
+							max={ 6 }
+						/>
+					) }
+					<SelectControl
+						className={ 'salary-data-table-picker__select' }
+						label={ __( 'Select Job Data source' ) }
+						value={ queryTable }
+						options={ getQueryTables() }
+						onChange={ toggleAttribute( 'queryTable' ) }
+					/>
+					<TextControl
+						label={ __( 'Linked Salary Data URL' ) }
+						help={ __(
+							'The full URL to a page with a corresponding Salary Data block to link to. Leave blank to link to the current page.'
+						) }
+						value={ salaryDataUrl }
+						onChange={ toggleAttribute( 'salaryDataUrl' ) }
+					/>
+				</PanelBody>
+			</InspectorControls>
+			{ ! queryTable && (
 				<Placeholder
 					icon="admin-post"
 					label={ __( 'Job Classification Data' ) }
@@ -92,39 +232,20 @@ function JobClassifications( { attributes, setAttributes, tables } ) {
 						)
 					) }
 				</Placeholder>
-			</>
-		);
-	}
-
-	return (
-		<>
-			{ inspectorControls }
-			<ServerSideRender
-				block="hrswpsqlsrv/job-classifications"
-				attributes={ attributes }
-			/>
-		</>
+			) }
+			{ queryTable && isRequesting && (
+				<Placeholder
+					icon="admin-post"
+					label={ __( 'Job Classification Data' ) }
+				>
+					<Spinner />
+				</Placeholder>
+			) }
+			{ ! isRequesting &&
+				jobClassificationData?.length > 0 &&
+				( displayAsList
+					? renderJobClassificationList()
+					: renderJobClassificationTable() ) }
+		</div>
 	);
 }
-
-export default withSelect( ( select ) => {
-	const allTables = select( 'hrswpsqlsrv/salary-data' ).getTableNames();
-	let tables;
-
-	if ( Array.isArray( allTables ) ) {
-		tables = allTables.reduce( ( accumulator, currentValue ) => {
-			if (
-				currentValue.value.includes( 'job-class' ) ||
-				'' === currentValue.value
-			) {
-				accumulator.push( currentValue );
-			}
-
-			return accumulator;
-		}, [] );
-	}
-
-	return {
-		tables,
-	};
-} )( JobClassifications );