+++ src/blocks/job-classifications/index.js
@@ -1,35 +1,16 @@
 /**
- * WordPress dependencies
- */
-const { __, _x } = wp.i18n;
-
-/**
  * Internal dependencies
  */
 import edit from './edit';
 import metadata from './block.json';
-import save from './save';
 import icon from './icon';
 
-const { name, category, supports } = metadata;
+const { name } = metadata;
 
-export { name };
+export { metadata, name };
 
 export const settings = {
-	title: __( 'HRS Job Classifications' ),
-	category,
-	description: __( 'Display WSU job classification data.' ),
 	icon,
-	keywords: [ __( 'table data compensation jobs' ) ],
-	supports,
-	styles: [
-		{
-			name: 'default',
-			label: _x( 'Table', 'block style' ),
-			isDefault: true,
-		},
-		{ name: 'list', label: _x( 'List', 'block style' ) },
-	],
+	example: {},
 	edit,
-	save,
 };