+++ src/store/index.js
@@ -6,6 +6,8 @@ const { registerStore } = wp.data;
 
 const DEFAULT_STATE = {
 	tableNames: {},
+	jobClassificationData: {},
+	salaryData: {},
 };
 
 const actions = {
@@ -13,6 +15,18 @@ const actions = {
 		return {
 			type: 'GET_TABLE_NAMES',
 			tableNames,
+		};
+	},
+	getJobClassificationData( jobClassificationData ) {
+		return {
+			type: 'GET_JOB_CLASSIFICATION_DATA',
+			jobClassificationData,
+		};
+	},
+	getSalaryData( salaryData ) {
+		return {
+			type: 'GET_SALARY_DATA',
+			salaryData,
 		};
 	},
 	fetchFromAPI( path ) {
@@ -32,6 +46,16 @@ export default function registerStores() {
 						...state,
 						tableNames: action.tableNames,
 					};
+				case 'GET_JOB_CLASSIFICATION_DATA':
+					return {
+						...state,
+						jobClassificationData: action.jobClassificationData,
+					};
+				case 'GET_SALARY_DATA':
+					return {
+						...state,
+						salaryData: action.salaryData,
+					};
 			}
 
 			return state;
@@ -42,8 +66,15 @@ export default function registerStores() {
 		selectors: {
 			getTableNames( state ) {
 				const { tableNames } = state;
-
 				return tableNames;
+			},
+			getJobClassificationData( state ) {
+				const { jobClassificationData } = state;
+				return jobClassificationData;
+			},
+			getSalaryData( state ) {
+				const { salaryData } = state;
+				return salaryData;
 			},
 		},
 
@@ -58,6 +89,20 @@ export default function registerStores() {
 				const path = '/hrswp-sqlsrv-db/v1/tables/';
 				const tableNames = yield actions.fetchFromAPI( path );
 				return actions.getTableNames( tableNames );
+			},
+			*getJobClassificationData( table ) {
+				const path = `/hrswp-sqlsrv-db/v1/jobclassification/table/${ table }`;
+				const jobClassificationData = yield actions.fetchFromAPI(
+					path
+				);
+				return actions.getJobClassificationData(
+					jobClassificationData
+				);
+			},
+			*getSalaryData( table ) {
+				const path = `/hrswp-sqlsrv-db/v1/salarydata/table/${ table }`;
+				const salaryData = yield actions.fetchFromAPI( path );
+				return actions.getSalaryData( salaryData );
 			},
 		},
 	} );