+++ CHANGELOG.md
@@ -17,6 +17,23 @@ Changelog formatting (http://semver.org/):
 ### Security (in case of vulnerabilities.)
 -->
 
+## 0.10.0 (2022-06-06)
+
+### Added
+
+- Add WP Rest API endpoint and store for salary data. (b1eb163)
+- Add rest route for Job Classification data. (f3bb06b)
+
+### Changed
+
+- Bump WP tested-to to 6.0.0.
+- Convert salary data block to API v2 and stop using `serverSideRender`. (b1eb163)
+- Refactor Job Classifications block to use API v2 and replace `ServerSideRender`. (f3bb06b)
+- Switch to using the HRS theme global style variables. (1f7db38)
+- Add GitHub API to plugin header for updates. (510192d)
+- Modify WebPack config to copy `block.json` files to the `build` directory for block registration. (f3bb06b)
+- Bump nanoid from 3.1.30 to 3.2.0. (34fe424)
+
 ## 0.9.0 (2021-12-23)
 
 ### Changed